+++ angular-workspace/projects/ni/nimble-angular/table-column/enum-text/nimble-table-column-enum-text.directive.ts
@@ -1,0 +1,80 @@
+import { Directive, ElementRef, Input, Renderer2 } from '@angular/core';
+import { type TableColumnEnumText, tableColumnEnumTextTag } from '@ni/nimble-components/dist/esm/table-column/enum-text';
+import { BooleanValueOrAttribute, NumberValueOrAttribute, toBooleanProperty, toNullableNumberProperty } from '@ni/nimble-angular/internal-utilities';
+import { NimbleTableColumnBaseDirective } from '@ni/nimble-angular/table-column';
+import type { MappingKeyType } from '@ni/nimble-components/dist/esm/table-column/enum-base/types';
+
+export type { TableColumnEnumText, MappingKeyType };
+export { tableColumnEnumTextTag };
+
+/**
+ * Directive to provide Angular integration for the table column element for enum text.
+ */
+@Directive({
+    selector: 'nimble-table-column-enum-text'
+})
+export class NimbleTableColumnEnumTextDirective extends NimbleTableColumnBaseDirective<TableColumnEnumText> {
+    public get fieldName(): string | undefined {
+        return this.elementRef.nativeElement.fieldName;
+    }
+
+    // Renaming because property should have camel casing, but attribute should not
+    // eslint-disable-next-line @angular-eslint/no-input-rename
+    @Input('field-name') public set fieldName(value: string | undefined) {
+        this.renderer.setProperty(this.elementRef.nativeElement, 'fieldName', value);
+    }
+
+    public get keyType(): MappingKeyType {
+        return this.elementRef.nativeElement.keyType;
+    }
+
+    // Renaming because property should have camel casing, but attribute should not
+    // eslint-disable-next-line @angular-eslint/no-input-rename
+    @Input('key-type') public set keyType(value: MappingKeyType) {
+        this.renderer.setProperty(this.elementRef.nativeElement, 'keyType', value);
+    }
+
+    public get fractionalWidth(): number | null | undefined {
+        return this.elementRef.nativeElement.fractionalWidth;
+    }
+
+    // Renaming because property should have camel casing, but attribute should not
+    // eslint-disable-next-line @angular-eslint/no-input-rename
+    @Input('fractional-width') public set fractionalWidth(value: NumberValueOrAttribute | null | undefined) {
+        this.renderer.setProperty(this.elementRef.nativeElement, 'fractionalWidth', toNullableNumberProperty(value));
+    }
+
+    public get minPixelWidth(): number | null | undefined {
+        return this.elementRef.nativeElement.minPixelWidth;
+    }
+
+    // Renaming because property should have camel casing, but attribute should not
+    // eslint-disable-next-line @angular-eslint/no-input-rename
+    @Input('min-pixel-width') public set minPixelWidth(value: NumberValueOrAttribute | null | undefined) {
+        this.renderer.setProperty(this.elementRef.nativeElement, 'minPixelWidth', toNullableNumberProperty(value));
+    }
+
+    public get groupIndex(): number | null | undefined {
+        return this.elementRef.nativeElement.groupIndex;
+    }
+
+    // Renaming because property should have camel casing, but attribute should not
+    // eslint-disable-next-line @angular-eslint/no-input-rename
+    @Input('group-index') public set groupIndex(value: NumberValueOrAttribute | null | undefined) {
+        this.renderer.setProperty(this.elementRef.nativeElement, 'groupIndex', toNullableNumberProperty(value));
+    }
+
+    public get groupingDisabled(): boolean {
+        return this.elementRef.nativeElement.groupingDisabled;
+    }
+
+    // Renaming because property should have camel casing, but attribute should not
+    // eslint-disable-next-line @angular-eslint/no-input-rename
+    @Input('grouping-disabled') public set groupingDisabled(value: BooleanValueOrAttribute) {
+        this.renderer.setProperty(this.elementRef.nativeElement, 'groupingDisabled', toBooleanProperty(value));
+    }
+
+    public constructor(renderer: Renderer2, elementRef: ElementRef<TableColumnEnumText>) {
+        super(renderer, elementRef);
+    }
+}