+++ angular-workspace/projects/ni/nimble-angular/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log - @ni/nimble-angular
 
-This log was last generated on Mon, 14 Aug 2023 14:35:00 GMT and should not be manually modified.
+This log was last generated on Mon, 14 Aug 2023 16:14:55 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 17.0.16
+
+Mon, 14 Aug 2023 16:14:55 GMT
+
+### Patches
+
+- Angular support for enum-text table column ([ni/nimble@a8e29e5](https://github.com/ni/nimble/commit/a8e29e5042ddb9ad9258b09c4844f5c360844628))
 
 ## 17.0.15